Seperate curriculums and courses into different files
