+++ routes/router.go
@@ -2,6 +2,7 @@ package routes
 
 import (
 	"tat_gogogo/controllers/curriculum"
+	//"tat_gogogo/controllers/curriculum/courses"
 	"tat_gogogo/controllers/login"
 
 	"github.com/gin-gonic/gin"
@@ -11,8 +12,7 @@ import (
 RegisterRoutes is a place to register rotes
 */
 func RegisterRoutes(router *gin.Engine) {
-
 	router.POST("/login", login.Controller)
 	router.POST("/curriculums", curriculum.Controller)
-	router.POST("/curriculums/course", curriculum.CourseController)
+	router.POST("/curriculums/courses", curriculum.CoursesController)
 }
